backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
pydantic==1.10.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.0
redis==4.5.0
pyjwt==2.6.0
